requirements.txt
add-trailing-comma==2.0.1
aiosqlite==0.12.0
appdirs==1.4.3
aspy.refactor-imports==2.1.0
asyncpg==0.20.1
atomicwrites==1.4.0
attrs==19.3.0
autoflake==1.3.1
bandit==1.6.2
black==19.10b0
cached-property==1.5.1
cfgv==3.1.0
ciso8601==2.1.3
click==7.1.1
colorama==0.4.3
dead==1.3.0
distlib==0.3.0
entrypoints==0.3
eradicate==1.0
filelock==3.0.12
flake8==3.7.9
flake8-annotations==2.1.0
flake8-annotations-coverage==0.0.3
flake8-bandit==2.1.2
flake8-broken-line==0.2.0
flake8-bugbear==20.1.4
flake8-builtins==1.5.2
flake8-commas==2.0.0
flake8-comprehensions==3.2.2
flake8-debugger==3.2.1
flake8-eradicate==0.3.0
flake8-executable==2.0.3
flake8-fixme==1.1.1
flake8-future-import==0.4.6
flake8-mutable==1.2.0
flake8-plugin-utils==1.3.0
flake8-polyfill==1.0.2
flake8-pyi==19.3.0
flake8-pytest==1.3
flake8-pytest-style==1.1.1
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-unused-arguments==0.0.3
gitdb==4.0.4
GitPython==3.1.1
h11==0.9.0
httptools==0.1.1
identify==1.4.15
iso8601==0.1.12
itsdangerous==1.1.0
mccabe==0.6.1
more-itertools==8.2.0
nodeenv==1.3.5
packaging==20.3
passlib==1.7.2
pathspec==0.8.0
pbr==5.4.5
pluggy==0.13.1
pre-commit==2.3.0
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPika==0.37.1
pytest==5.4.2
pyupgrade==2.3.0
PyYAML==5.3.1
regex==2020.4.4
reorder-python-imports==2.2.0
simple-print==0.1
six==1.14.0
smmap==3.0.2
starlette==0.13.3
stevedore==1.32.0
termcolor==1.1.0
tokenize-rt==4.0.0
toml==0.10.0
tortoise-orm==0.16.8
typed-ast==1.4.1
typing-extensions==3.7.4.2
uvicorn==0.11.3
virtualenv==20.0.18
wcwidth==0.1.9
websockets==8.1
yesqa==1.1.0
